Use the EOL constant for line endings
This makes the testsuite (`npm test`) pass for me on GNU/Linux
which would previously fail due to unexpected newline characters
in the test.

I think the line ending used in the GetFeature-XML is `\r\n` in the repo,
which get converted to `\n` on GNU/Linux once the repository is cloned.

There might be more clever fixes, but this should make the test run
both on windows and on GNU/Linux.