Interpreter directive must be the first line
This commit moves the "hashbang" to the first line so that the "-e"
parameter is honoured.

On most systems the "#!/bin/bash -e" would have been ignored, and
instead the script was likely being interpreted by /bin/sh directly.
A way to test this is to try something like the following:

$ cat > subvert-shebang.sh <<EOF
> #This comment on line 1 subverts the interpreter directive on line 2
> #!/bin/bash -e
> echo 1
> false
> echo "oh no, this should not happen"
> EOF
$ chmod +x ./subvert-shebang.sh
$ ./subvert-shebang.sh
1
oh no, this should not happen
$

The text "starting with a line" is present, yet not emphasized, in
the "man 2 execve" man page:

       execve() executes the program pointed to by filename.

       filename  must  be  either a binary executable, or a script
       starting with a line of the form:

           #! interpreter [optional-arg]