MPR-102 add control component for table style
The TableStyleControl is supposed to be used in the ?style=table mode. It shares its props signature with the ControlIcon and can be used interchangeably to render a control as table tool.

In a next step, it has to be used in FullScreen, OverviewMap, and TotalView when table style is detected as soon as their respective MPR-102 branches are merged.