add MRH-30: vue draw tool modify features
The draw tool (vue) is now able to change values of features (like color and size).

Set to state "modify" to select a feature via single click (if you hold
the click, it will change the shape as usual).
Edit the feature with the select- and input-boxes provided by the draw tool.

Every feature will now hold the state of selected values from the store
for the time the feature was created (see addDrawStateToFeature). This
way the values of a feature to modify are preselected in select- and
input-boxes.

Also kml, geojson and gpx imports are recognized by the drawtool and can
be manipulated as stated above.

New listener: createSelectInteractionModifyListener
New store value: selectedFeature