MRH-109 Fix - close the externe Fachdaten in menu
If you load an external wms and open the "externe Fachdaten" in the menu,
loading a second wms, it is only shown in "externe Fachdaten" after closing and reopening "externe Fachdaten".

The solution is to close "externe Fachdaten" in the menu every time a external wms is loaded.
Also fixed the bug for the function closeAllExpandedFolder in core
modellist file.