fix MRH 3d Layer - legend error
when we switch the map from 2d to 3d and click the button i of the
layers under 3d data, sometime it shows the error in the console that
this.createLegend is not a function. Because the layer type terrain has
not the function createLegend.

Here we test first if this.createLegend function is there or not, if it
is there, the function will be executed. And the function createLegend
is also implemented for the layer type terrain.