fix MRH layerInformation: ie11 error
Only the first click in ie11 on an "i" to open the layerInformation
will trigger a legend to be shown. Any other click on any "i" will
have no legend content in ie11.

The layerInformation is rendered by Vue (id=legend) but than put
into the Backbone layerInformation (id=layerinformation-desktop).

Following this adventurous approach, we have to wait with this.$nextTick
and then clone the outerHTML from the Vue and hide the Vue legend.

This is a dirty solution, but we can't think of any other fix.