add MPAIS-11 draw tool: config double circle
The contour color from both circles of the double circle can now be
configured in different colors. It is also possible to edit the colors
of both circles with one drop down box seperately.

The one thing that doesn't work is to recognize the outer color and set
the drop down box correctly.

New config param: outerColorContour