MPAIS-85 bug: drawtool colortailing
Drawing something (e.g. a symbol) with the draw tool and
switching colors in-between will lead formerly drawn elements
to switch their colors as well.

We need a clone of styleSettings each time a draw event is called
otherwise the copy will influence former drawn objects.

This is the bugfix.