MPAIS-91 Fix: the invisible feature will not be printed
The feature which is from draw tool or tacktical tool could be set
visible or invisible by the filter. If the feature is invisible, there
will be error by printing.

The goal is to make the invisible feature not printable and the
customers just can print what they see in the map.