MPAIS-94 Fix: make the download button enabled after the draw tool reopens
If the user draws some features and save the features, after closing
the reopening the draw tool and also insert the new name and new format,
the download button is still disabled even there are drawn features from
last manipulation.

The problem is that the features for download is empty after the user reopens
the draw tool, now by setting the download feature with drawn features
from previous steps to make the download button enabled.