G31DEV1-62 Refactor: Layer Id config.json could also be as Object
Before the layer id in config.json is only as string like id: "123",
every id represents a layer. But for project Bildungsatlas, we need to
filter features from the same layer to set them as 'Layers' in menu
tree.

With the id in an Object format we could set the same layer id and different
suffix text to make different themes from the same layer but
show as different layers in the menu tree.

Here another function is also implemented: For WFS layer, we could set
PROPERTYNAME in config.json to filter the response data from WFS layer
to avoid load all the data from Server.