MPVERKEHR-197 change layer order after folder click
By click on a folder in the topics list ("Themenbaum") will activate
its layers. The order layers are activated is willingly inverted,
leading layers that are configured to be on top of others to be
behind them in the map once you click on the folder.

This commit makes the former fixed inverted order configurable,
introducing the "invertLayerOrder" flag for folders.