MPAIS-104 Fix: Making text feature seem closer to original feature in KML file
Up to now in the draw tool if the user draws a text and export it as kml
file, after reimporting the file, there are some problems: 1. Besides
text there are always a yellow pin, this is because that there are no
styles for the feature in kml file, especially no icon for this feature. 2. The text has no style so that it
shows a little bitte strange.

To solve these problems, 1. Setting a fake icon as style for the text.
2. After reading about KML, it is impossible to make the style 1:1 as in
draw tool, the color can be delivered, but the text fonts style could
not, and in kml file, the only parameter to resize the text is "scale",
in the code the scale will be calculated with font size.