MRH 7 Fix: Making external wms layers of which the version is lower than 1.3 can also be checked by the extent and imported.
Before mrh only allows the external wms layers with version 1.3.0 or above to be imported. If the wms layers with version lower than 1.3.0 is imported, an error message will show.

The big difference is in new version the parameter CRS replaced of SRS which is in old version to present the format of boundingbox. In masterportal the openlayer can only parse the parameter CRS.

Now by importing the wms layers with lower version, all parameter SRS will be changed into CRS so that materportal can parse it in order to check if the layer is in mrh boundingbox.

Another function here is to make a green tip for successful import and a red cross for failed import.