G31DEV1-111 Add: Getting data with wfs filter in config.json
For WFS layer, it could be used filter to get specific data without
loading all the data. The filter is written in xml format and saved in a
directory and in config.json it is needed to add the file path with the
parameter "wfsFilter". Then if a wfs layer with filter is activated, the
wfs layer parsing function in masterportal will parse the data.

Before it used ajax call to get data from service manager and now it is
replaced with axios.